feat: factor out cpuidpage to a shared crate
