+++ crates/shared/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "shared"
+version = "0.7.0"
+edition = "2021"
+description = "API for the Enarx hypervisor-microkernel boundary"
+repository = "https://github.com/enarx/enarx"
+license = "Apache-2.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+const-default = { workspace = true, features = ["derive"] }
