+++ crates/shared/src/no_std/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//! share code between the Enarx components not using std
+
+pub mod cpuid_page;
