+++ crates/shim-kvm/src/snp/cpuid_page.rs
@@ -1,108 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! FIXME: move to sev crate
+//! Structures and methods to handle the SEV-SNP CPUID page
 
 use core::arch::x86_64::CpuidResult;
-use core::fmt::Debug;
-use core::mem::size_of;
-
-use const_default::ConstDefault;
 
 use crate::snp::snp_active;
 use crate::_ENARX_CPUID;
-
-const COUNT_MAX: usize = 64;
-
-/// An entry in the SNP CPUID Page
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Default, ConstDefault, Eq, PartialEq)]
-pub struct CpuidFunctionEntry {
-    /// function
-    pub eax_in: u32,
-    /// index
-    pub ecx_in: u32,
-    /// register state when cpuid is called
-    pub xcr0_in: u64,
-    /// register state when cpuid is called
-    pub xss_in: u64,
-    /// cpuid out
-    pub eax: u32,
-    /// cpuid out
-    pub ebx: u32,
-    /// cpuid out
-    pub ecx: u32,
-    /// cpuid out
-    pub edx: u32,
-    reserved: u64,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug)]
-struct CpuidPageEntry {
-    count: u32,
-    reserved_1: u32,
-    reserved_2: u64,
-    functions: [CpuidFunctionEntry; COUNT_MAX],
-}
-
-impl ConstDefault for CpuidPageEntry {
-    const DEFAULT: Self = CpuidPageEntry {
-        count: 0,
-        reserved_1: 0,
-        reserved_2: 0,
-        functions: [<CpuidFunctionEntry as ConstDefault>::DEFAULT; COUNT_MAX],
-    };
-}
-
-impl Default for CpuidPageEntry {
-    fn default() -> Self {
-        <Self as ConstDefault>::DEFAULT
-    }
-}
-
-/// The CPUID page to be copied in the guest VM
-#[repr(C)]
-#[derive(Copy, Clone, Debug)]
-pub struct CpuidPage {
-    entry: CpuidPageEntry,
-    space: [u8; CpuidPage::space_size()],
-}
-
-impl CpuidPage {
-    #[allow(clippy::integer_arithmetic)]
-    const fn space_size() -> usize {
-        4096 - size_of::<CpuidPageEntry>()
-    }
-
-    /// Get all entries
-    #[inline]
-    pub fn get_functions(&self) -> &[CpuidFunctionEntry] {
-        &self.entry.functions[..self.entry.count as usize]
-    }
-}
-
-impl ConstDefault for CpuidPage {
-    const DEFAULT: Self = CpuidPage {
-        entry: ConstDefault::DEFAULT,
-        space: [0; CpuidPage::space_size()],
-    };
-}
-
-impl Default for CpuidPage {
-    fn default() -> Self {
-        <Self as ConstDefault>::DEFAULT
-    }
-}
-
-/// Error thrown by CpuidPage methods
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum Error {
-    /// no memory allocating structs for the ioctl
-    NoMemory,
-    /// the page already contains the maximum number of entries
-    Full,
-}
 
 /// See [`cpuid_count`](cpuid_count).
 #[inline]
